flare-relayer-main/project/requirements/remote.txt
-r common.txt

# server
uWSGI==2.0.23

# Prometheus Monitoring
prometheus-client==0.19.0
django-prometheus==2.3.1

# Sentry Monitoring
sentry-sdk==1.39.2

# Unittest Reporting
unittest-xml-reporting==3.2.0
coverage==7.4.0
